// EscalafonFrontend/src/app/interfaces/provincia.ts
export interface Provincia {
    /**
     * @type { number } - iPrvnId
     */
    iPrvnId?: number;

    /**
     * @type { number } - iDptoId
     */
    iDptoId?: number;

    /**
     * @type { string } - cPrvnCodigo
     */
    cPrvnCodigo?: string;

    /**
     * @type { string } - cPrvnNombre
     */
    cPrvnNombre?: string;

    /**
     * @type { string } - cPrvnAbreviatura
     */
    cPrvnAbreviatura?: string;

    /**
     * @type { number } - iPrvnEstado
     */
    iPrvnEstado?: number;

    /**
     * @type { string } - cPrvnGentilicio
     */
    cPrvnGentilicio?: string;

}