// EscalafonFrontend/src/app/interfaces/distrito.ts
export interface Distrito {
    /**
     * @type { number } - iDsttId
     */
    iDsttId?: number;

    /**
     * @type { number } - iPrvnId
     */
    iPrvnId?: number;

    /**
     * @type { string } - cDsttCodigo
     */
    cDsttCodigo?: string;

    /**
     * @type { string } - cDsttNombre
     */
    cDsttNombre?: string;

    /**
     * @type { string } - cDsttAbreviatura
     */
    cDsttAbreviatura?: string;

    /**
     * @type { number } - iDsttEstado
     */
    iDsttEstado?: number;

    /**
     * @type { string } - cDsttGentilicio
     */
    cDsttGentilicio?: string;

}