// EscalafonFrontend/src/app/interfaces/infopefamiliar-domicilio.ts
export interface InfopefamiliarDomicilio {
    /**
     * @type { number } - iSecInfoPeFamDomiId
     */
    iSecInfoPeFamDomiId?: number;

    /**
     * @type { number } - iLegId
     */
    iLegId?: number;

    /**
     * @type { number } - iTipDirId
     */
    iTipDirId?: number;

    /**
     * @type { number } - iZonaId
     */
    iZonaId?: number;

    /**
     * @type { number } - iTipViaId
     */
    iTipViaId?: number;

    /**
     * @type { number } - iDptoId
     */
    iDptoId?: number;

    /**
     * @type { number } - iPrvnId
     */
    iPrvnId?: number;

    /**
     * @type { number } - iDsttId
     */
    iDsttId?: number;

    /**
     * @type { string } - cInfoPeFamDireccion
     */
    cInfoPeFamDireccion?: string;

    /**
     * @type { string } - cInfoPeFamReferencia
     */
    cInfoPeFamReferencia?: string;

    /**
     * @type { number } - iArchId
     */
    iArchId?: number;

}